CRUD 모두 백단내에서는 AUTH 미들웨어 쿠키토큰  권한이 있어야 할수 있게 함

+++ server/controller/snippetController.js
@@ -65,6 +65,9 @@ exports.updateOneSnippet = async (req, res) => {
       });
     }
 
+    if (originalSnippet.user.toString() !== req.user)
+      return res.status(401).json({ errorMessage: "수정권한이 없습니다." });
+
     originalSnippet.title = title;
     originalSnippet.description = description;
     originalSnippet.code = code;
@@ -95,6 +98,9 @@ exports.deleteOneSnippet = async (req, res) => {
         errorMessage: "No snippet ID",
       });
     }
+
+    if (existingSnippet.user.toString() !== req.user)
+      return res.status(401).json({ errorMessage: "삭제권한이 없습니다." });
 
     await existingSnippet.delete();
     res.json(existingSnippet);
